@@ -1,4 +1,4 @@
-package looper
+package lockerredis
 
 import (
 	"context"
@@ -8,6 +8,8 @@ import (
 	"github.com/go-redsync/redsync/v4"
 	"github.com/go-redsync/redsync/v4/redis/goredis/v9"
 	"github.com/redis/go-redis/v9"
+
+	"github.com/golang-cz/looper"
 )
 
 // redisOptions := &redis.Options{
@@ -23,10 +25,10 @@ import (
 
 // RedisLocker provides an implementation of the Locker interface using
 // redis for storage.
-func RedisLocker(ctx context.Context, rc redis.UniversalClient) (locker, error) {
+func RedisLocker(ctx context.Context, rc redis.UniversalClient) (looper.Locker, error) {
 	err := rc.Ping(ctx).Err()
 	if err != nil {
-		return nil, fmt.Errorf("%s: %w", ErrFailedToConnectToLocker, err)
+		return nil, fmt.Errorf("%s: %w", looper.ErrFailedToConnectToLocker, err)
 	}
 
 	pool := goredis.NewPool(rc)
@@ -38,13 +40,13 @@ func RedisLocker(ctx context.Context, rc redis.UniversalClient) (locker, error) 
 }
 
 // Locker
-var _ locker = (*redisLocker)(nil)
+var _ looper.Locker = (*redisLocker)(nil)
 
 type redisLocker struct {
 	rs *redsync.Redsync
 }
 
-func (r *redisLocker) lock(ctx context.Context, key string, timeout time.Duration) (lock, error) {
+func (r *redisLocker) Lock(ctx context.Context, key string, timeout time.Duration) (looper.Lock, error) {
 	options := []redsync.Option{
 		redsync.WithTries(1),
 		redsync.WithExpiry(timeout + time.Second),
@@ -52,7 +54,7 @@ func (r *redisLocker) lock(ctx context.Context, key string, timeout time.Duratio
 	mu := r.rs.NewMutex(key, options...)
 	err := mu.LockContext(ctx)
 	if err != nil {
-		return nil, ErrFailedToObtainLock
+		return nil, looper.ErrFailedToObtainLock
 	}
 
 	rl := &redisLock{
@@ -63,20 +65,20 @@ func (r *redisLocker) lock(ctx context.Context, key string, timeout time.Duratio
 }
 
 // Lock
-var _ lock = (*redisLock)(nil)
+var _ looper.Lock = (*redisLock)(nil)
 
 type redisLock struct {
 	mu *redsync.Mutex
 }
 
-func (r *redisLock) unlock(ctx context.Context) error {
+func (r *redisLock) Unlock(ctx context.Context) error {
 	unlocked, err := r.mu.UnlockContext(ctx)
 	if err != nil {
-		return ErrFailedToReleaseLock
+		return looper.ErrFailedToReleaseLock
 	}
 
 	if !unlocked {
-		return ErrFailedToReleaseLock
+		return looper.ErrFailedToReleaseLock
 	}
 
 	return nil
